Rename socketTypes to socket_types
