+++ app/javascript/projects/modelling/socket_types.ts
